requirements-test.txt
# Testing Dependencies for AgriGuard CI/CD Pipeline

# Core testing framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
coverage==7.3.2

# HTTP testing
httpx==0.25.1
requests-mock==1.11.0

# FastAPI testing
fastapi==0.104.1
uvicorn==0.24.0

# Mock and fixtures
pytest-mock==3.12.0
faker==20.0.3

# Code quality
flake8==6.1.0
black==23.11.0
mypy==1.7.0

# Application dependencies (needed for tests)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
google-cloud-storage==2.10.0
google-generativeai==0.3.1
chromadb==0.4.18
